Raise EVDutyApiInvalidCredentialsError only when creds are invalid
The EVduty backend only supports 1 access token per account.

The token this API uses is invalidated when another access token is
created by another app (for example, the android app).

Since it can happen a lot, this API only clears the invalidated token,
so the next time a call will be made, a new token will be created.